Phase 2a: Extract db subcommand into commands/db.py
Moves run_db_command into its own module. Imports tightened to only
what db administration actually needs (db_admin helpers, settings,
stats backend resolvers).

The standalone subcommand splits (build, resume, db) are now complete.
Remaining Phase 2a work: split the 'tasks' and 'stats' dispatchers
into their own packages with one module per sub-action.

commands/__init__.py shrinks below 1500 lines.
